Move date calculation functions to utils module
